Moved Expr to utils to avoid circular imports
